unsafeLazyMMapVectors: Decrease chunk size to ease memory pressure
When the chunks are very large, there is very little room to map them
properly, and even a small delay in reclaiming the vectors can prevent
the next chunk from being mapped leading to out of memory errors.

By decreasing the size of the chunks they become easier to map,
because they fit in more open spaces in the programs memory, and they
are less sensitive to delays in garbage collection because more than
one chunk can fit in the address space simultaneously.

The downside is that it limits the size of the vectors that can be
mapped. However, the current limit of 2^24 bytes per vector on a
32-bit architecture, is probably enough for most uses.